+++ package.json
@@ -68,7 +68,7 @@
   },
   "dependencies": {
     "@aws-sdk/client-dynamodb": "3.651.1",
-    "@govuk-one-login/di-ipv-cri-common-express": "10.3.0",
+    "@govuk-one-login/di-ipv-cri-common-express": "10.6.0",
     "@govuk-one-login/frontend-analytics": "3.0.0",
     "@govuk-one-login/frontend-language-toggle": "1.1.0",
     "@govuk-one-login/frontend-passthrough-headers": "1.1.1",
